+++ server/src/main/kotlin/net/horizonsend/ion/server/IonServer.kt
@@ -16,8 +16,8 @@ import net.horizonsend.ion.server.miscellaneous.handle
 import net.horizonsend.ion.server.miscellaneous.initializeCrafting
 import net.horizonsend.ion.server.miscellaneous.listeners
 import net.minecraft.core.registries.BuiltInRegistries
-import net.horizonsend.ion.server.generation.SpaceBiomeProvider
-import net.horizonsend.ion.server.generation.SpaceChunkGenerator
+import net.horizonsend.ion.server.features.generation.generators.SpaceBiomeProvider
+import net.horizonsend.ion.server.features.generation.generators.SpaceChunkGenerator
 import net.starlegacy.feature.economy.city.CityNPCs
 import net.starlegacy.feature.economy.collectors.Collectors
 import net.starlegacy.feature.hyperspace.HyperspaceBeacons
