+++ server/src/main/kotlin/net/horizonsend/ion/server/features/generation/generators/SpaceChunkGenerator.kt
@@ -1,4 +1,4 @@
-package net.horizonsend.ion.server.generation
+package net.horizonsend.ion.server.features.generation.generators
 
 import org.bukkit.World
 import org.bukkit.generator.BlockPopulator
